refactor(IntegrationTests): location
moved tests that expect ExitCode.Success to the SuccessPrograms folder
